Configuration: Added two new types: auth+coa and auth+acct+coa.
For listen section added CoA retransmission config same way it is done
for home_server. So that TCP server is able to send/retransmit CoA
requests with some pattern.

For home_server section, coa subsection added a virtual_server option to
be able to process coa messages in unlung config.