Store and find.
Once I have a 'reverse' listener I need to be able to store that. That is done either by key (TCP-Session-Key) or by
address. An address means just an IP here. Note I store a list in the tree, so that several listeners may share the same
key or IP address. If this is the case and several listeners share the same key or IP address then I select a next
listener to be used (to send CoA request) in a round-robin fashion. So that if switch has several tunnels (and they all
are exactly equal for us) we will evenly split the load between them. For IP use case I have added an option to
explicitly specify port number, in which case it is possible to select
exact listener without any round-robin.