Store listener by Address and TCP-Session-Key
I am able to store the listener by address as soon as I have it reversed (because I know an IP right away). At the same
time, I need to wait for a special attribute (and special packet optionally (Status-Server)) to be able to store reverse
listener by its key.

I need to add internal attribute here as well so that to be able to configure the key.