Originating a CoA
First the original flow tries to find a home_server where to send a CoA. It is done either by IP(port) or by
Home-Server/Home-Pool reference. If that failed we are trying to find a reverse tunnel to send the CoA to. That is done
either by IP and optionally port or by a key that may be set from TCP-Session-Key earlier. If such reverse tunnel is
found we may get a home_server from this reversed listener. Home-Server has been previously created when we have
reversed the listener.

From now on original flow continues as if the home_server has been found with the old approach. The difference is in
insert_into_proxy_hash function where we are trying to allocate an ID for the request *and* to setup a proxy listener to
be used to actually send the CoA. Since we already have known this proxy listener to be used (this is our 'reversed'
listener) we don't need to setup it. We need just to allocate an ID.

The limitation here is that we *must* use a specific tunnel only. Hence, we cannot send more than 256 CoA request at the
same time. In the original flow freeRADIUS may open another connection if all IDs are allocated for a particular
listener/connection. We have no such option within this feature since by definition we *must* use single tunnel to send
CoAs.

Potential workaround is the case when *switch* opens several tunnels. If these tunnels are equal and may
process all CoAs interchangeably, then if all IDs have been already allocated for the current reversed listener we may
try another reverse listener if available.

In the end of the flow a send function pointer is called. This send callback is different for the reversed listener
compared to the original listener. This is specified in the next commit. Stay tuned.