Send/recv for reversed listeners.
TCP Server (RAD_LISTEN_AUTH type) had a pair of functions
    -) dual_tls_recv - receives requests
    -) dual_tls_send - sends *responses*

So we need to make sure we may *send* requests as well. For that a new funtion is introduced: dual_tls_send_req.

For recv part - we simply cannot differentiate a packet until we actually read it, so the receiving is always happening
on the original listener. Once a packet is received we may check its type and if required (for CoA/Disconnect process it with the reversed
listener.

The similar logic applies for TCP Client (RAD_LISTEN_PROXY). So it had a pair:
    -) proxy_tls_recv - receives responses
    -) proxy_tls_send - sends *requests*

So new function is introduced to send *responses* instead of requests: proxy_tls_send_reply.

At the same time original proxy_tls_recv should be able to receive not just responses but also requests. If
CoA/Disconnect request packet type is detected we run the request flow with the reversed listener. Not that in that case
(since we are working with the reversed listener) at the time the flow ends and ready to send a "response", a new
version of send is called: proxy_tls_send_reply.