Lifecycle management.
Each listener has its own lifecycle enum state assigned. Now we have a reversed listener. We need to make sure the
states of these two are in sync. That also means that if some actions are required to be done as lifecycle state is
changed similar actions should be made for both main(original) and reversed listeners (check event_new_fd function).